feat: add IMauiView for automatic view model injection
